libhardware_legacy/wifi: add workaround for ar6000 driver
The ar6000 driver requires the interface to be brought up in order
for scanning to function. Without this change, the supplicant fails
on first activation of wifi, and will only connect successfully
after multiple retries. This workaround is also necessary to fix
Wi-Fi hotspot mode (in conjunction with another change to
system/netd).

Backported from CM7.

Change-Id: I11fa55ab23295d40041df097450b73484f52b042